+++ unique_sdk/CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.70] - 2026-01-13
+- Adding additional parameters `isQueueable`, `executionOptions` and `progressTitle` to the message execution
+
 ## [0.10.69] - 2026-01-16
 - Add unified type checking CI with basedpyright